// src/app/shared/enum/kham-benh.enum.ts
export enum LoaiHangDoi {
    ChuanBiKham = 1,
    LamChiDinh = 2,
    DoiKetLuan = 3
}

export enum TrangThaiHangDoi {
    ChoKham = 1,
    DangKham = 2
}

export enum TrangThaiYeuCauDichVuKyThuat {
    //ChuaDongTien = 0,
    ChuaThucHien = 1,
    DangThucHien = 2,
    DaThucHien = 3,
    DaHuy = 4
}

export enum TrangThaiYeuCauKhamBenh {
    ChuaKham = 1,
    DangKham = 2,
    DangLamChiDinh = 3,
    DangDoiKetLuan = 4,
    DaKham = 5,
    HuyKham = 6
}

export enum EnumNhomGoiDichVu {
    DichVuKhamBenh = 1,
    DichVuKyThuat = 2,
    VatTuTieuHao = 3,
    DuocPham = 4,
    DichVuGiuongBenh = 5,
    DonThuocThanhToan = 6
}

export enum EnumTypeLoadingData {
    View = 1,
    Update = 2,
    Delete = 3,
    Cancel = 4,
    Add = 5
}

export enum EnumTypeComponent {
    KInput = 1,
    KDropdown = 2,
    KTextBox = 3
}
export enum LoaiDichVuKyThuat {
    ThuThuatPhauThuat = 1,
    XetNghiem = 2,
    ChuanDoanHinhAnh = 3,
    ThamDoChucNang = 4,
    TheoYeuCau = 5,
    Khac = 6
}
export enum LoaiPhieuIn {
    InPhieuChiDinh = 1,
    InDonThuocBHYT = 2,
    InDonThuocNhaThuoc = 3,
    InPhieuKhamBenh = 4,
    InPhieuKhamBenhVaoVien = 5,
}

export enum EnumTrangThaiYeuCauTiepNhan {
    DangThucHien = 1,
    DaHoanTat = 2,
    DaHuy = 3
}