// src/app/shared/enum/tiem-vacxin.enum.ts
export enum LoaiKetLuanKhamSangLocTiemChung {
    DuDieuKienTiem = 1,
    TamHoanTiemChung = 2,
    KhongDongYTiem = 3,
    ChongChiDinh = 4
}

export enum LoaiPhanUngSauTiem {
    KhongCoPhanUng = 1,
    PhanUngThongThuong = 2,
    TaiBienNang = 3,
    PhanUngKhac = 4
}

export enum ViTriTiem {
    TiemBap = 1,
    TiemTinhMach = 2,
    TiemDuoiDa = 3,
    TiemTrongDa = 4,
    Uong = 5
}

export enum TrangThaiTiemChung {
    ChuaTiemChung = 1,
    DaTiemChung = 2,
    // ChongChiDinh = 3,
    // KhongDongYTiem = 4
}

export enum LaiHangDoiTiemVacxin {
    KhamSangLoc = 1,
    ThucHienTiem = 2,
    LichSuTiem = 3
}

export enum LoaiTienSuTiemVacxin {
    BanThan = 1,
    GiaDinh = 2
}

export enum LoaiDiUngTiemVacxin {
    Thuoc = 1,
    ThucAn = 2,
    Khac = 3
}

export enum MucDoDiUngTiemVacxin {
    //Xác nhận lại DS enum
    Nang = 1
}

export enum NoiXuTriTheoDoiTiemVacxin {
    TaiNha = 1,
    TaiTramYTe = 2,
    BenhVienTuyenHuyen = 3,
    BenhVienTuyenTinhTrungUong = 4,
    YTeTuNhan = 5,
    Khac = 6
}

export enum TinhTrangHienTaiTheoDoiTiemVacxin {
    Khoi = 1,
    DiChung = 2,
    TuVong = 3,
    Khac = 4
}

export enum EnumTrangThaiHangDoi {
    ChoKham = 1,
    DangKham = 2
}